fix(console): one component per file, and a selected row that says so
AGENTS.md:166 asks for one JSX-returning component per `.tsx`, and
KnowledgeIcon.tsx held two. Split into KnowledgeTypeIcon and
KnowledgeSectionIcon, which is also where a reader would look for either.

The navigator's selection was visual only — `data-selected` and a heavier
weight. `aria-current` on all three row types gives a screen reader the same
signal a sighted reader gets from the highlight.

Both raised by CodeRabbit on the previous head.
