+++ packages/console/src/components/knowledge/KnowledgeSectionIcon.tsx
@@ -16,21 +16,6 @@ import {
 } from 'lucide-react'
 
 /**
- * One glyph per `type:`, so a drawer of a dozen notes can be read by shape
- * instead of by reading every title. The profile's own vocabulary — a slice is a
- * buildable piece, an entity a thing the app is about, a decision a rule that
- * rules something out — and anything a project invents falls back to a page.
- */
-const TYPE_ICONS: Record<string, LucideIcon> = {
-  overview: BookOpen,
-  slice: Layers,
-  entity: Boxes,
-  decision: Scale,
-  question: HelpCircle,
-  note: FileText,
-}
-
-/**
  * One glyph per section, matched on the last segment so `decisions/security`
  * gets the lock rather than the scales its parent wears. A section the profile
  * does not name — anything a project adds — takes the icon of the section it
@@ -47,14 +32,6 @@ const SECTION_ICONS: Record<string, LucideIcon> = {
   personas: Users,
   actors: UserRound,
   log: ScrollText,
-}
-
-export const KnowledgeTypeIcon: React.FC<{
-  type?: string
-  size?: number
-}> = ({ type, size = 13 }) => {
-  const Icon = (type && TYPE_ICONS[type]) || FileText
-  return <Icon size={size} color="var(--mantine-color-dimmed)" />
 }
 
 export const KnowledgeSectionIcon: React.FC<{
